requirements.txt
Flask==3.1.2
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
flask-cors==5.0.0

pdf2image==1.17.0
pdfplumber==0.11.4
pytesseract==0.3.10
Pillow==10.2.0

python-docx==1.1.0
python-pptx==0.6.21
openpyxl==3.1.2

openai==1.13.3
PyMuPDF==1.24.5
pypdf==4.2.0
gunicorn
